+++ edit-me/types/CMSLink.d.ts
@@ -1,0 +1,7 @@
+import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
+
+interface CMSLink {
+  href: string;
+  icon: IconDefinition;
+  title: string;
+}