+++ src/components/Articles/AboutMe.tsx
@@ -1,18 +1,14 @@
+import { personal } from '@content';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
 import React from 'react';
-import { CMSPersonalInformation } from '../../cms-integration/markdown/personal';
 import Prose from '../Prose/Prose';
 import { SectionHeading } from '../SectionHeading/SectionHeading';
 
-interface AboutMeProps {
-  personalInformation: CMSPersonalInformation;
-}
-
-const AboutMe: React.FC<AboutMeProps> = ({ personalInformation }) => {
+const AboutMe: React.FC = () => {
   return (
     <article>
       <SectionHeading icon={faUser} level={3} text="About Me" />
-      <Prose html={personalInformation.html} />
+      <Prose html={personal.body.html} />
     </article>
   );
 };