+++ src/components/Articles/ContactInformation.tsx
@@ -1,16 +1,13 @@
+import { PrivateField, personal } from '@content';
 import { faIdCard } from '@fortawesome/free-solid-svg-icons';
 import React from 'react';
-import { CMSPersonalInformation } from '../../cms-integration/markdown/personal';
-import { CMSPrivateInformation } from '../../cms-integration/markdown/private';
 import { SectionHeading } from '../SectionHeading/SectionHeading';
 
 interface ContactInformationProps {
-  personalInformation: CMSPersonalInformation;
-  privateInformation?: CMSPrivateInformation[];
+  privateInformation?: PrivateField[];
 }
 
 export const ContactInformation: React.FC<ContactInformationProps> = ({
-  personalInformation,
   privateInformation,
 }) => {
   return (
@@ -19,14 +16,14 @@ export const ContactInformation: React.FC<ContactInformationProps> = ({
 
       <ul className="mt-2">
         <li>
-          <strong>Location:</strong> {personalInformation.attributes.location}
+          <strong>Location:</strong> {personal.location}
         </li>
 
         {/* private access required */}
         {privateInformation?.map((privateField) => (
-          <li className="mt-3" key={privateField.attributes.label}>
-            <strong>{privateField.attributes.label}</strong>{' '}
-            <div dangerouslySetInnerHTML={{ __html: privateField.html }} />
+          <li className="mt-3" key={privateField.label}>
+            <strong>{privateField.label}</strong>{' '}
+            <div dangerouslySetInnerHTML={{ __html: privateField.body.html }} />
           </li>
         ))}
       </ul>