+++ src/components/Prose/Prose.tsx
@@ -1,11 +1,14 @@
+import { twMerge } from 'tailwind-merge';
+
 interface ProseProps {
+  className?: string;
   html: string;
 }
 
-const Prose: React.FC<ProseProps> = ({ html }) => {
+const Prose: React.FC<ProseProps> = ({ className, html }) => {
   return (
     <div
-      className="prose mt-2 text-base text-neutral-12"
+      className={twMerge('prose mt-2 text-base text-neutral-12', className)}
       dangerouslySetInnerHTML={{ __html: html }}
     />
   );