+++ src/pages/api/pdf.tsx
@@ -1,15 +1,13 @@
+import { allPrivateFields } from '@content';
 import ReactPDF from '@react-pdf/renderer';
 import { NextApiHandler } from 'next';
-import { getCMSIntegration } from '../../cms-integration/getCMSIntegration';
-import { getPrivateInformation } from '../../cms-integration/markdown/private';
 import PDF from '../../components/PDF/PDF';
 
 const privateKey = process.env.PRIVATE_KEY;
 
 const handler: NextApiHandler = async (req, res) => {
-  const props = await getCMSIntegration('markdown');
-
   const secret = req.query.secret;
+
   let privateInformation;
   if (typeof secret !== 'undefined') {
     if (secret !== privateKey) {
@@ -17,11 +15,11 @@ const handler: NextApiHandler = async (req, res) => {
       res.end('Not authorized');
       return null;
     }
-    privateInformation = await getPrivateInformation();
+    privateInformation = allPrivateFields;
   }
 
   const pdfStream = await ReactPDF.renderToStream(
-    <PDF privateInformation={privateInformation} {...props} />,
+    <PDF privateInformation={privateInformation} />,
   );
   res.setHeader('Content-Type', 'application/pdf');
   pdfStream.pipe(res);